+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 4.16.2 Jun 27, 2021
+
+Upgrade priority: Low. Recommended for Polkadot with runtime >= 9050.
+
+Changes:
+
+- Correct `CompactAssignments` & `RawSolution` types for Polkadot
+
+
 ## 4.16.1 Jun 26, 2021
 
 Upgrade priority: Low. Recommended for full from-genesis queries on Kusama.